feat: pedidos observaciones changes

+++ src/main/java/ar/com/trazabilidad/webservice/PedidosRS.java
@@ -1,8 +1,10 @@
 package ar.com.trazabilidad.webservice;
 
+import ar.com.trazabilidad.dominio.Observaciones;
 import ar.com.trazabilidad.dominio.PedidoDetalle;
 import ar.com.trazabilidad.dominio.Pedidos;
 import ar.com.trazabilidad.dominio.Productos;
+import ar.com.trazabilidad.servicio.ObservacionesService;
 import ar.com.trazabilidad.servicio.PedidosService;
 import ar.com.trazabilidad.servicio.ProductosService;
 import java.util.List;
@@ -23,22 +25,25 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
 @RestController
+@CrossOrigin(origins = "http://localhost:5173")
 @RequestMapping("/pedidos")
-@CrossOrigin
 @Slf4j
 public class PedidosRS {
-    
+
     @Autowired
     PedidosService service;
-    
+
     @Autowired
     ProductosService productosService;
-    
+
+    @Autowired
+    ObservacionesService observacionService;
+
     @GetMapping("/")
     public List<Pedidos> obtener() {
         return service.findAll();
     }
-    
+
     @GetMapping("/id/{id}")
     public ResponseEntity obtenerPorId(@PathVariable("id") Integer id) {
         Optional<Pedidos> pedido = service.findById(id);
@@ -48,35 +53,36 @@ public class PedidosRS {
             );
         }
         return ResponseEntity.of(pedido);
-        
+
     }
-    
+
     @PostMapping("/")
     public Pedidos crearPedido(@RequestBody Pedidos pedido) {
         return service.save(pedido);
-        
+
     }
-    
+
     @PostMapping("/terminado")
     public Pedidos editarPedidoTerminado(@RequestBody Pedidos pedido) {
-        
+
         for (PedidoDetalle pedidoDetalle : pedido.getPedidoDetalleList()) {
-            
+
             Productos producto = pedidoDetalle.getIdproducto();
-            
+
             producto.setCantidad(producto.getCantidad() - pedidoDetalle.getCantidad().intValue());
-            
+
             productosService.save(producto);
-            
+
         }
         return service.save(pedido);
     }
-    
+
     @PostMapping("/all")
     public List<Pedidos> crearPedidos(@RequestBody List<Pedidos> pedidos) {
         return service.saveAll(pedidos);
     }
-    
+
+    @DeleteMapping("/id/{id}")
     public ResponseEntity borrarPorId(@PathVariable("id") Integer id) {
         Optional<Pedidos> pedido = service.findById(id);
         if (!pedido.isPresent()) {
@@ -84,7 +90,15 @@ public class PedidosRS {
                     HttpStatus.NOT_FOUND, "No se encontro un pedido para borrar con el ID proporcionado"
             );
         }
+        List<Observaciones> obsAsociadas = observacionService.findByIdpedido(pedido.get());
+
+        if (!obsAsociadas.isEmpty()) {
+            throw new ResponseStatusException(
+                    HttpStatus.CONFLICT, "No se encontro un pedido para borrar con el ID proporcionado"
+            );
+        }
         service.deleteById(id);
         return ResponseEntity.ok("Se borro correctamente");
     }
+
 }
